latest version

+++ devoir2/Unishop/src/Data/Entities/Users/User.java
@@ -56,7 +56,7 @@ public class User {
         return number;
     }
 
-    public void setNumber(int number) {
+    public void setNumber(long number) {
         this.number = number;
     }
